refactor: 즐겨 찾기 업데이트 하는 함수 모듈화
